requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2024.2.2
decorator==5.1.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.12.0
pillow==10.2.0
py==1.11.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
retry==0.9.2
setuptools==68.2.2
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
wheel==0.41.2
zhipuai==2.0.1
